feat: feature/ResourceModal resource 정보 보내기 기능 구현 및 controller 분리
