+++ models/Resource.js
@@ -2,7 +2,10 @@ const mongoose = require("mongoose");
 
 const resourceSchema = new mongoose.Schema({
   name: { type: String, required: true, unique: true },
-  categoryId: String,
+  categoryId: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: "Category",
+  },
   currentVersion: {
     type: mongoose.Schema.Types.ObjectId,
     ref: "ResourceVersion",
