+++ routes/categories.js
@@ -3,9 +3,11 @@ const router = express.Router();
 const {
   resource,
   getResourceList,
+  getResource,
 } = require("../controllers/categories.controller");
 
-router.post("/:categoryId/resource", resource);
 router.get("/:categoryId", getResourceList);
+router.post("/:categoryId/resource", resource);
+router.get("/:categoryId/resources/:resourceId", getResource);
 
 module.exports = router;
